feat(forge create): read calldata from file (#674)
* feat(forge create): read calldata from file

It is possible to create a transaction where the
calldata is too large to be passed via `argv`.
See https://wiki.debian.org/CommonErrorMessages/ArgumentListTooLong
for more information on this limitation. To make it
possible to create these large transactions, update
`forge create` so that it can read the calldata
from a file.

A new cli argument is added `--constructor-args-path`
that conflicts with `--constructor-args`. This new
argument should point to a file on the filesystem
that contains a whitespace delimited list of
constructor arguments.

* refactor out `mut`

Co-authored-by: Georgios Konstantopoulos <me@gakonst.com>

* Update cli/src/cmd/create.rs

Co-authored-by: Georgios Konstantopoulos <me@gakonst.com>

* fix: get compiling

* chore: fmt with latest nightly

* chore: clippy lints

Co-authored-by: Georgios Konstantopoulos <me@gakonst.com>