feat: cast wallet generation - json output and many wallets generation (#6160)
* Add multiple wallet creation and JSON output support

This change extends wallet creation functionality to allow the generation of multiple wallets. It introduces a new CLI option --number that allows the user to specify the number of wallets to generate. Additionally, a new --json flag is added that provides the option for users to output wallet information in JSON format. This feature will be especially useful for scripts or automated processes that need to interact with the wallet. It improves usability and flexibility of the wallet creation process.

* Improve keystore creation success message

* fix

* Update mod.rs

Co-authored-by: evalir <hi@enriqueortiz.dev>

* Update mod.rs

Co-authored-by: evalir <hi@enriqueortiz.dev>

* Fix

---------

Co-authored-by: auebasota1337 <105765683+auebasota1337@users.noreply.github.com>
Co-authored-by: evalir <hi@enriqueortiz.dev>